Cut redundant dataset I/O per sample and per startup scan
Sample lookup globbed both the image and mask directories on every __getitem__,
so each fetch cost two full directory scans and an epoch was quadratic in the
number of files -- costly on a parallel filesystem. Resolve the id-to-path
mapping once at construction (asserting one file per stem) and look it up in
constant time per item.

Masks are shipped from the loader in a narrow int16 carrier rather than widened
to int64 in the worker; the trainer and evaluator already cast to long on the
compute device, so this quarters the pinned-memory footprint and host-to-device
transfer for masks. In the non-sharded path the volume-preparation copy is now
skipped when np.load already returned a fresh, contiguous, correctly-typed array
(the materializing copy is kept only for the mmap-backed sharded path). A
mask-only accessor lets class-weight estimation read masks without also loading
and preprocessing the full image volume.

+++ tests/test_infra.py
@@ -85,11 +85,12 @@ def test_tiny_dataset_v2_loads(tiny_dataset):
     assert train.dataset_format_version == 2
 
     item = train[0]
-    # Channels-first volume, float; single-channel long mask.
+    # Channels-first volume, float; single-channel narrow-int mask carrier
+    # (widened to long on the compute device by the trainer/evaluator).
     assert tuple(item["image"].shape) == (3, 16, 16, 16)
     assert item["image"].dtype == torch.float32
     assert tuple(item["mask"].shape) == (16, 16, 16)
-    assert item["mask"].dtype == torch.int64
+    assert item["mask"].dtype == torch.int16
     assert int(item["mask"].max()) <= 2
 
     # The raw on-disk arrays match the format the loader/trainer expect.
